Moves Java JdbcRDD test case to a separate test suite
